fix: move schemas to eval_schemas.py
